# swagger.yaml
swagger: '2.0'
info:
  description: This is the web version of the libopac Java library
  version: 0.1.0
  title: libopac-web-api
  contact:
    email: opac@k-nut.eu
host: berlin.codefor.de
basePath: /
tags:
  - name: media_data
    description: Information about things to borrow from the libraries
    
schemes:
  - https
  
paths:
  '/libraries/{libraryName}/search':
    get:
      tags:
        - media_data
      summary: Search for media - libopac style
      description: Search for media - structured after libopac
      produces:
        - application/json
      parameters:
        - name: libraryName
          in: path
          description: ID of the library
          required: true
          type: string
        - name: title
          in: query
          description: Title
          required: false
          type: string
        - name: author
          in: query
          description: Author
          required: false
          type: string
        - name: isbn
          in: query
          description: ISBN
          required: false
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/RawResult'
        '400':
          description: Invalid tag value
  '/libraries/{libraryName}/{isbn}':
    get:
      tags:
        - media_data
      summary: Retrieve information by ISBN10 or ISBN13
      description:  Record Details Endpoint structured after libraryextension
      produces:
        - application/json
      parameters:
        - name: libraryName
          in: path
          description: ID of the library
          required: true
          type: string
        - name: isbn
          in: path
          description: ISBN of the book
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/RecordDetails'
        '400':
          description: Invalid input
  '/libraries/{libraryName}/catalog/search':
    get:
      tags:
        - media_data
      summary: Search for media - libraryextension style
      description: Search for media - structured after libraryextension
      produces:
        - application/json
      parameters:
        - name: libraryName
          in: path
          description: ID of the library
          required: true
          type: string
        - name: searchTerm
          in: query
          description: The Name of the Field in which to search
          required: true
          type: string
          enum:
            - ISBN
            - ISBN13
            - Title
        - name: term
          in: query
          description: The string to search for
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/StructuredResponse'
        '400':
          description: Invalid input
  '/libraries/{libraryName}/record/{id}':
    get:
      tags:
        - media_data
      summary: Retrieve book by id - libraryextension style
      produces:
        - application/json
      parameters:
        - name: libraryName
          in: path
          description: ID of the library
          required: true
          type: string
        - name: id
          in: path
          description: ID of the media
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Details'
        '400':
          description: Invalid input
  '/libraries/{libraryName}/record/branches/{id}':
    get:
      tags:
        - media_data
      summary: Availability Endpoint
      produces:
        - application/json
      parameters:
      - name: libraryName
        in: path
        description: ID of the library
        required: true
        type: string
      - name: id
        in: path
        description: ID of the media
        required: true
        type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Availability'
        '400':
          description: Invalid input

          
definitions:
  RawResult:
    type: object
    properties:
      type:
        type: string
      nr:
        type: number
      id:
        type: string
      innerhtml:
        type: string
      status:
        type: string
        enum:
          - GREEN
          - RED
      coverBitmap:
        type: string
      page:
        type: number
      childquery:
        type: string

    
  StructuredResponse:
    type: object
    properties:
      id:
        type: string
      infoHtml:
        type: string
      format:
        type: string

  Details:
    type: object
    properties:
      details:
          $ref: '#/definitions/Details1'
  Details1:
    type: object
    properties:
      total_copies:
        type: integer
        example: 1
      available_copies:
        type: integer
        example: 1
      branches:
        type: array
        items:
          type: string
          example: "ZB Grimm-Zentrum"
  Availability:
    type: object
    properties:
      branches:
        type: array
        items:
          type: string
          example: "ZB Grimm-Zentrum"
  RecordDetail:
    type: object
    properties:
      description:
        type: string
        example: <b>Der Weg des Bogens</b><br />Coelho, Paulo<br />2017
      availability:
        type: string
        example: RESTRICTED
  RecordDetails:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: '#/definitions/RecordDetail'